requirements.txt
requests==2.31.0
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.29.0